+++ src/gql/types/PhoneNumberType.php
@@ -8,6 +8,7 @@
 namespace rynpsc\phonenumber\gql\types;
 
 use craft\gql\GqlEntityRegistry;
+use craft\helpers\ArrayHelper;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type;
 
@@ -20,17 +21,11 @@ use GraphQL\Type\Definition\Type;
  */
 class PhoneNumberType
 {
-    /**
-     * @return string
-     */
     public static function getName(): string
     {
         return 'phoneNumber_PhoneNumber';
     }
 
-    /**
-     * @return Type
-     */
     public static function getType(): Type
     {
         if ($type = GqlEntityRegistry::getEntity(self::class)) {
@@ -44,35 +39,88 @@ class PhoneNumberType
         ]));
     }
 
-    /**
-     * @return array
-     */
     public static function getFieldDefinitions(): array
     {
         return [
-            'region' => [
-                'name' => 'region',
+            'carrierName' => [
+                'name' => 'carrierName',
                 'type' => Type::string(),
+                'args' => ['locale' => Type::string()],
+                'resolve' => function($source, $arguments) {
+                    return $source->getCarrierName(ArrayHelper::getValue($arguments, 'locale'));
+                },
+            ],
+            'countryCode' => [
+                'name' => 'countryCode',
+                'type' => Type::string(),
+            ],
+            'description' => [
+                'name' => 'description',
+                'type' => Type::string(),
+                'args' => [
+                    'locale' => Type::string(),
+                    'region' => Type::string(),
+                ],
+                'resolve' => function($source, $arguments) {
+                    return $source->getDescription(
+                        ArrayHelper::getValue($arguments, 'locale'),
+                        ArrayHelper::getValue($arguments, 'region'),
+                    );
+                },
+            ],
+            'extension' => [
+                'name' => 'extension',
+                'type' => Type::string(),
+            ],
+            'format' => [
+                'name' => 'format',
+                'type' => Type::string(),
+                'args' => ['format' => Type::string()],
+                'resolve' => function($source, $arguments) {
+                    return $source->format(ArrayHelper::getValue($arguments, 'format'));
+                },
+            ],
+            'formatCountry' => [
+                'name' => 'formatForCountry',
+                'type' => Type::string(),
+                'args' => ['region' => Type::string()],
+                'resolve' => function($source, $arguments) {
+                    return $source->formatForCountry(ArrayHelper::getValue($arguments, 'region'));
+                },
+            ],
+            'formatMobileDialing' => [
+                'name' => 'formatForMobileDialing',
+                'type' => Type::string(),
+                'args' => [
+                    'region' => Type::string(),
+                    'format' => Type::boolean(),
+                ],
+                'resolve' => function($source, $arguments) {
+                    return $source->formatForMobileDialing(
+                        ArrayHelper::getValue($arguments, 'region'),
+                        ArrayHelper::getValue($arguments, 'format'),
+                    );
+                },
             ],
             'number' => [
                 'name' => 'number',
                 'type' => Type::string(),
             ],
-            'countryCode' => [
-                'name' => 'countryCode',
+            'region' => [
+                'name' => 'region',
                 'type' => Type::string(),
             ],
             'regionCode' => [
                 'name' => 'regionCode',
                 'type' => Type::string(),
             ],
+            'timezones' => [
+                'name' => 'timezones',
+                'type' => Type::listOf(Type::string()),
+            ],
             'type' => [
                 'name' => 'type',
                 'type' => Type::int(),
-            ],
-            'description' => [
-                'name' => 'description',
-                'type' => Type::string(),
             ],
         ];
     }