+++ src/helpers/StringHelper.php
@@ -16,20 +16,17 @@ namespace rynpsc\phonenumber\helpers;
  */
 class StringHelper extends \craft\helpers\StringHelper
 {
-    // Public Methods
-    // =========================================================================
-
     /**
-     * Multi-byte sub string replacement.
+     * Multibyte safe sub string replacement.
      *
-     * @param string $string
-     * @param mixed $replacement
-     * @param mixed $start
-     * @param mixed $length
+     * @param string $string The string to replace
+     * @param string $replacement The replacement value
+     * @param int $start
+     * @param int $length
      * @return string
      */
     public static function substrReplace(string $string, string $replacement, int $start, int $length): string
     {
-        return mb_substr($string, 0, $start).$replacement.mb_substr($string, $start + $length);
+        return mb_substr($string, 0, $start) . $replacement . mb_substr($string, $start + $length);
     }
 }