+++ src/translations/en/phone-number.php
@@ -6,8 +6,8 @@
  */
 
 return [
-    'Region' => 'Region',
-    'The number provided is invalid.' => 'The number provided is invalid.',
     'No results' => 'No results',
-    'Phone Number' => 'Phone Number'
+    'Phone Number' => 'Phone Number',
+    'Region' => 'Region',
+    'The phone number provided is invalid.' => 'The phone number provided is invalid.',
 ];